BugFix: restore correct handling to SIGNALS in pigpio
Prior code attempted to assign a generic handler to catch ALL signals;
however SIGKILL and SIGSTOP cannot be caught; also the default action for
SIGTTOU, SIGTTIN, SIGTSTP and SIGCONT should perhaps be "ignored" NOT
"terminate", these signals are associated with Job Control and the first
three WILL halt execution until receipt of the forth SIGCONT.  As a result
the previous code will result in program termination when it is run in a
debugger such as gdb and the program is halted at a breakpoint - which is
not helpful when debugging!

The previous signal handler uses a macro "DBG(...)" however that uses
the printf(...) family of functions - **this is not safe as the signal
handler can be executed asynchronously (at any time and at any point in the
main program flow) and printf() is not re-entrant!** Indeed a big warning
ought to be put into the documentation for setgpioSetSignalFunc(...) and
gpioSetSignalFuncEx(...) that the user provided function **must be safe to
run asynchronously** - for guidance it should ONLY contain functions in the
"safe-functions" list in POSIX.1-2004 (also known as POSIX.1-2001 Technical
Corrigendum 2) which is reproduced in the signal(7) man page under the
section headed "Async-signal-safe functions" with some revisions from
POSIX.1-2008 .

The range of signals that CAN be handled ranges from SIGHUP (1) to SIGRTMAX
however SIGRTMAX is NOT a compile time constant but a runtime macro that
typically expands to 64.  There is also a gap between SIGUNUSED (a.k.a.
SIGSYS, typically 31) and SIGRTMIN (either 34 or 35 depending on glibc
version and the real-time extensions system which uses two or three
"reserved" signals internally) the man page for signal(7) explicitly
states:
"Because the range of available real-time signals varies according to the
 glibc threading implementation (and this variation can occur at run time
 according to the available kernel and glibc), and indeed the range of real
 -time signals varies across UNIX systems, programs should never refer to
 real-time signals using hard-coded numbers, but instead should always
 refer to real-time signals using the notation SIGRTMIN+n, and include
 suitable (run-time) checks that SIGRTMIN+n does not exceed SIGRTMAX."

Because the signal handler can be executed whilst the signal handler for a
particular signal is being changed it is necessary to protect against it
using something that may be being changed in the main program loop.  The
fix that this commit proposes is to maintain a shadow copy of the array of
handlers and to use an atomic flag variable to toggle between using the
main and shadow copy.

This commit restores a default "ignore" action for those signals that
typically are so treated.  For the mentioned ones here a short debugging
message is also created that warns of the data lose that will occur from
pausing the execution of the pigpio library code.  This typically happens
if a breakpoint is reached in the debugger or if <Cntl>-z is pressed during
the running of a executable.  This commit also reverts to dumping core for
signals that do that by default - this does it by restoring the default
signal action for an otherwise unhandled signal in the generic handler that
is supposed to do so and then returns to the point in the code that invokes
it - a second signal should then occur which WILL dump the core and
terminate. This is a change to previous behaviour in that NO attempt is
made to clean up {by running gpioTerminate() or exit(...)} - which would
possibly corrupt the data that the dump would otherwise contain (as well as
moving the program location away from the point where the signal was raised
for!)

For otherwise unhandled signals that do not terminate pigpio the problems
with the DBG macro and its use of printf type functions is overcome by
using individual atomic flags for each such signal - this does mean that
a helper function (void)processSignals() is required. This MUST be included
in the main loop of pigpiod and other applications using the pigpio library
to produce the actions that were previously **unsafely** done in the signal
handler code.  This is the debug level adjustments caused by SIGUSR1 and
SIGUSR2 and the debug messages that most of the others produced - those
have been extended to include the extra signals this commit recognises as
falling in this category.  To prevent "undefined behaviour" it is necessary
to change away from using sleep() in the main wait loop of pigpiod as
it is subject to that issue if SIGALARM handling is modified which we allow
fortunately nanosleep() does not suffer from the same problem.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>